attack_sinusoid_deconstructor.py works
Transposing by changing the filter centres doesn't work that well, you get
bell-like tones.

This might be because the error signal is not white.

Also, the gain is not compensated for in designing the resonant filters, so most
likely some "partials" are too loud when resynthesizing with anything other than
the error signal (using the error signal will always get you the original
signal, basically we want the error to contain only "noise" information and the
resonances to contain all the pitch information).