We can get transposition to work OK by using a large hop size
Hop size has to be at least the window size it seems.

But shouldn't we be able to get it to work using a smaller hop? I guess it's
okay because that means less computation.

We cannot, however, get the residual signal very well. Not sure why this is. Is
it misalignment? Or some scaling problem?

Finally: maybe its better to build up spectra in the frequency domain using the
peak indices and using this to get the residual...