Tested qualitatively the time stretching with locked points
If you tweak the settings a bit, you can get it to sound good, i.e., it
preserves the attack portion and just stretches the tonal part.

We should still check how good the time synchronization is. If no critical
locked input points are discarded, the mapping should be correct, but I think an
offset is introduced because the time stretching part doesn't take into
consideration the start time of the first output frame (it just starts writing
at sample 0, increasing the frame time by H). So this should be checked.

Also we should try it out on more percussive sounds. Already, time stretching
the piano by a lot sounds unrealistic because the amplitude time constant is
also stretched. This can sound good, but for more realistic stretching, we might
also want to extrapolate the amplitude envelope.

Finally, we need a way to reliably detect the attacks so that the creation of
mapping points be automated.