Routines for datastructures of native implementation of pvoc_synth
We use KISSFFT because the FFTW interface is too inflexible.

What they say in the documentation about typically always transforming the same
arrays over and over again is true (they require you pass the arrays you will
transform in and out of when you intialize the DFT structures). But no other
library (e.g., ARM CMSIS, Intel IPP) that I can see uses this interface and in
the future when we port to an embedded system we will probably use the CMSIS
library. What is even more strange is that you have to use their fftw_malloc
routines to allocate the arrays you will transform, so why can't the library
just choose optimal addresses if they are so concerned about the addresses of
the input and output arrays?