It might be that pvoc_synth_f32_test works
But gives NaNs for pathological signals. This happens when the magnitude of a
bin is 0 which probably won't happen for real signals that have some noise
present.

Going to try some qualitative tests by stretching sine waves and the like. I
imagine that if a small amount of dither is added then this problem will not
happen.