Griffin and Lim were right
We're not dividing out the analysis and synthesis windows correctly.

The C implementation of the pvoc needs be fixed so it does this correctly.
Specifically the window needs to be divided out after all the ola has happened.

Also, why can't we have constant scaling at the beginning of the pvoc? I don't
want the first window to like 'fade in'.