Correcting phase from time-stretch by resetting
If we reset the phase by simply outputting an analysis window verbatim (without
computing the phase from the past), it sounds okay if you do this on a window
that is at least a window's length in front of the thing (e.g., an attack) whose
phase you want to preserve. I thought maybe you could get away with (window_size
- hop_size) before the attack, but maybe we hear the interference of the
verbatim window overlapping the previous recursively computed phase.

The next thing to do is try this with automatically computed "reset" times
(times when you use a verbatim window) and "reset" times estimated from the
signal.