Real-time pitch shift works for constant rate
The way we do it is cheesy, we just repeat the same analysis time until we have
enough samples. This analysis time is one window's length into the past (because
in real-time we have no access to future samples). This is the way that the
output timing agrees the most with the input. But it sounds okay.

Also right now there's no way to reset the phase when pitch-shifting in
real-time. A simple way might be to just do it when the input is quiet.