ADSR can now be used in blocks
Works not bad. The ADSR blocks new attacks during it's sustain section (they
start right after the sustain section if still asserted). This is not a bad way
because we probably want to fade to 0 to avoid clicks when starting a new note
or something, but might not be what everyone expects. So the alternative is to
pre-empt the attack, fading out quickly, then starting the attack shortly after,
but this sounds overly complicated, so not sure if I'll implement this mode.