We have to make when notes start and end consistent
ADSR indicates the end of a note having the last active sample of a note
non-zero, whereas the note_region_segmenter expects the first inactive sample to
be non-zero to indicate the end. This latter situation is more general because
start and ends can be generated from an arbitrary gate signal just by finding
first-order differences (and clipping both and negating the end-indicating
signal). So we will change ADSR to also output this (the end of the note is the
first zero sample).