Real-time pitchshifting avoid attacks is working ok
We had to increase the worst case ringbuffer size. The reason I think is
because the ringbuffer size is mis-calculated when using Pmin==Pmax to
calculate Nr and Nw, which is what we do in
test/pitch_shifting/ps_ts_file_const_multi_chan.py (we just pitch shift
by a constant). In that case it would exceed the worst case size, which
is bad. But increasing Nw by 1 in the QH calculation has so far not
triggered an error with the hundreds of random tests I performed.

As for the performance of the algorithm, on drums the sound is more
present than without attack compensation. If you put large windows and
hops, the time can get really distorted because it's sitting waiting for
a safe move for a long time.