Weighted spectral difference can help isolate the attacks
This seems to be because, in this case, it ignores the influence of
vibrato, whose amplitude of frequency modulation is more important in
the higher frequencies. This may help with percussive instruments like
guitar, because their upper harmonics decay more quickly, but still a
test should be done for this case.