Trying to smooth the amplitude tracks
Smoothing was attempted by determining an average amplitude "shape",
then applying it to each amplitude track by scaling it to that track's
range. This makes the sound very synthetic sounding, which maybe is cool
but not ideal.

Need to trying a different smoothing technique, such as finding the
local maxima at each frame in the amplitude spectrum of the partials and
"squishing" that maximum down if it exceeds too much its neighbours.