Time stretching is possible with hop partial tracking
This is done by interpolating the instantaneous frequencies and
re-accumulating the phases. It sounds not bad if the estimate of the
partials is good, but can amplify artifacts if the estimate of the
partials is poor. Also it will probably suffer from phenomena related to
partials that are actually measurements of noise, and it will stretch
the modulations too (it will not keep the vibrato rate, say).