we can indeed do the time-domain inner-product to find DFT dk
that means we don't need to take the fourier transform to find the
derivative w.r.t k in the frequency domain (duh, but I wanted to see it
with my own eyes).

Now it remains to be seen how well the gradient steps work with this
(they should work the same...).