on our way to adding time domain harmonic gradient computer
not finished

But also there have been thoughts surrounding determining if the f0
chosen is the correct octave.

The way to proceed is perhaps as follows: if we are looking at the
gradients of a harmonic signal and we are slightly too high in our
estimate for f0, then the gradients should fall off roughly linearly in
the log-domain (they would be linear in the log-power spectrum of a
harmonic signal windowed by a gaussian), because we will be estimating
the gradient further and further away from the peaks. If the octave we
have chosen for this pitch class is too low, the gradients will
oscillate between falling off linearly and being random-valued (take the
example where the assumed f0 is one octave too low: every other gradient
will be "random"). So somehow we need to choose f0 so that the fall-off
is the most linear / smooth. To do this we probably need access to each
gradient separately.